Create galaxy_+_function_main_model
this program models 2 different type of movements of the stars inside their own galaxies, and in the meantime draws a rotational velocity-distance function related to the stars. Inside one of the 2 galaxies I used one of the MOND (Modified Newtonian Dynamics) type of equation, and in the other galaxy, I used an equation based on the Kepler laws.